+++ jvm/src/main/scala/eval/tile/LazyRaster.scala
@@ -4,7 +4,7 @@ import com.azavea.maml.eval._
 
 import geotrellis.raster._
 import geotrellis.raster.mapalgebra.local._
-import geotrellis.raster.mapalgebra.focal.{Square, Neighborhood, TargetCell, Slope => GTFocalSlope}
+import geotrellis.raster.mapalgebra.focal.{Square, Neighborhood, TargetCell, Slope => GTFocalSlope, Aspect => GTAspect}
 import geotrellis.raster.mapalgebra.focal.hillshade.{Hillshade => GTHillshade}
 import geotrellis.raster.render._
 import geotrellis.vector.{Extent, MultiPolygon, Point}
@@ -198,6 +198,23 @@ object LazyRaster {
       GTHillshade(fst.evaluate, Square(1), gridbounds, cs, azimuth, altitude, zFactor, TargetCell.All)
     lazy val dblTile =
       GTHillshade(fst.evaluateDouble, Square(1), gridbounds, cs, azimuth, altitude, zFactor, TargetCell.All)
+
+    def get(col: Int, row: Int) = intTile.get(col, row)
+    def getDouble(col: Int, row: Int) = dblTile.get(col, row)
+  }
+
+  case class Aspect(
+    children: List[LazyRaster],
+    gridbounds: Option[GridBounds[Int]],
+    cs: CellSize
+  ) extends UnaryBranch {
+    override lazy val cols: Int = gridbounds.map(_.width).getOrElse(fst.cols)
+    override lazy val rows: Int = gridbounds.map(_.height).getOrElse(fst.rows)
+
+    lazy val intTile =
+      GTAspect(fst.evaluate, Square(1), gridbounds, cs, TargetCell.All)
+    lazy val dblTile =
+      GTAspect(fst.evaluateDouble, Square(1), gridbounds, cs, TargetCell.All)
 
     def get(col: Int, row: Int) = intTile.get(col, row)
     def getDouble(col: Int, row: Int) = dblTile.get(col, row)